Add basic product list.
Using ember.js to handle the page.  Not much use of it right now, but it
will become super handy.

In lieu of an actual api, just using a raw json file for now for testing
things out.  This will be replaced in the near future by an actual API.

There's a bit of a hack in the css so that the thumbnails will space out
correctly if they don't all fit on one row.  This particular method
doesn't work in older IE's, but that's good enough for me.