Move json loading and the chain to link to outlet.
Doing it this way should make this all much easier to extend as the data
models get more complicated and more numerous.  This is very similar to
the way it's done in travis-web (used a chunk of the code from its route
for builds).  This also opens up the ability to do something special
while the data is being fetched from the api.