Update to ember 1.0.0-rc.1 and use ember-data.
This is a huge change for the application but keeps it up to date and
switches it to use ember-data to access the API.  This will help
tremendously in adding new functionality and maintaining it.
Documentation for ember before 1.0 will run scarce soon so it's
definitely best to be up to date.  With ember data getting "official"
support as well as widespread community support, using it should also
greatly ease functionality.